metodo faltastock

+++ src/main/java/com/Bazar/Bazar/DTO/BazarDTO.java
@@ -11,14 +11,14 @@ public class BazarDTO {
     
     private Long codigo_venta;
     private Double total;
-    private Double cantidadProductos;
+    private int cantidadProductos;
     private String nombrecliente;
     private String ApellidoCliente;
 
     public BazarDTO() {
     }
 
-    public BazarDTO(Long codigo_venta, Double total, Double cantidadProductos, String nombrecliente, String ApellidoCliente) {
+    public BazarDTO(Long codigo_venta, Double total, int cantidadProductos, String nombrecliente, String ApellidoCliente) {
         this.codigo_venta = codigo_venta;
         this.total = total;
         this.cantidadProductos = cantidadProductos;
